Fixed bug with PhysicsSystem marking every movable entity as dirty every tick. Reduced network payload of empty states from 49B to 15B.

+++ Robust.Shared/Map/MapManager.Network.cs
@@ -65,6 +65,17 @@ namespace Robust.Shared.Map
                 grid => new GameStateMapData.GridCreationDatum(grid.ChunkSize, grid.SnapSize,
                     grid.IsDefaultGrid));
 
+            // no point sending empty collections
+            if (gridDatums.Count        == 0) gridDatums        = default;
+            if (gridDeletionsData.Count == 0) gridDeletionsData = default;
+            if (mapDeletionsData.Count  == 0) mapDeletionsData  = default;
+            if (mapCreations.Count      == 0) mapCreations      = default;
+            if (gridCreations.Count     == 0) gridCreations     = default;
+
+            // no point even creating an empty map state if no data
+            if (gridDatums == null && gridDeletionsData == null && mapDeletionsData == null && mapCreations == null && gridCreations == null)
+                return default;
+
             return new GameStateMapData(gridDatums, gridDeletionsData, mapDeletionsData, mapCreations, gridCreations);
         }
 
@@ -84,71 +95,80 @@ namespace Robust.Shared.Map
 
             // First we need to figure out all the NEW MAPS.
             // And make their default grids too.
-            foreach (var (mapId, gridId) in data.CreatedMaps)
+            if(data.CreatedMaps != null)
             {
-                if (_maps.ContainsKey(mapId))
+                foreach (var (mapId, gridId) in data.CreatedMaps)
                 {
-                    continue;
-                }
-                var gridCreation = data.CreatedGrids[gridId];
-                DebugTools.Assert(gridCreation.IsTheDefault);
+                    if (_maps.ContainsKey(mapId))
+                    {
+                        continue;
+                    }
+                    var gridCreation = data.CreatedGrids[gridId];
+                    DebugTools.Assert(gridCreation.IsTheDefault);
 
-                var newMap = new Map(this, mapId);
-                _maps.Add(mapId, newMap);
-                MapCreated?.Invoke(this, new MapEventArgs(newMap));
-                newMap.DefaultGrid = CreateGrid(newMap.Index, gridId, gridCreation.ChunkSize, gridCreation.SnapSize);
+                    var newMap = new Map(this, mapId);
+                    _maps.Add(mapId, newMap);
+                    MapCreated?.Invoke(this, new MapEventArgs(newMap));
+                    newMap.DefaultGrid = CreateGrid(newMap.Index, gridId, gridCreation.ChunkSize, gridCreation.SnapSize);
+                }
             }
 
             // Then make all the other grids.
-            foreach (var (gridId, creationDatum) in data.CreatedGrids)
+            if(data.CreatedGrids != null)
             {
-                if (creationDatum.IsTheDefault || _grids.ContainsKey(gridId))
+                foreach (var (gridId, creationDatum) in data.CreatedGrids)
                 {
-                    continue;
-                }
+                    if (creationDatum.IsTheDefault || _grids.ContainsKey(gridId))
+                    {
+                        continue;
+                    }
 
-                CreateGrid(data.GridData[gridId].Coordinates.MapId, gridId, creationDatum.ChunkSize,
-                    creationDatum.SnapSize);
+                    CreateGrid(data.GridData[gridId].Coordinates.MapId, gridId, creationDatum.ChunkSize,
+                        creationDatum.SnapSize);
+                }
             }
 
-            SuppressOnTileChanged = true;
-            // Ok good all the grids and maps exist now.
-            foreach (var (gridId, gridDatum) in data.GridData)
+            if(data.GridData != null)
             {
-                var grid = _grids[gridId];
-                if (grid.MapID != gridDatum.Coordinates.MapId)
+                SuppressOnTileChanged = true;
+                // Ok good all the grids and maps exist now.
+                foreach (var (gridId, gridDatum) in data.GridData)
                 {
-                    throw new NotImplementedException("Moving grids between maps is not yet implemented");
-                }
-
-                grid.WorldPosition = gridDatum.Coordinates.Position;
-
-                var modified = new List<(MapIndices position, Tile tile)>();
-                foreach (var chunkData in gridDatum.ChunkData)
-                {
-                    var chunk = grid.GetChunk(chunkData.Index);
-                    DebugTools.Assert(chunkData.TileData.Length == grid.ChunkSize * grid.ChunkSize);
-
-                    var counter = 0;
-                    for (ushort x = 0; x < grid.ChunkSize; x++)
-                    for (ushort y = 0; y < grid.ChunkSize; y++)
+                    var grid = _grids[gridId];
+                    if (grid.MapID != gridDatum.Coordinates.MapId)
                     {
-                        var tile = chunkData.TileData[counter++];
-                        if (chunk.GetTile(x, y).Tile != tile)
+                        throw new NotImplementedException("Moving grids between maps is not yet implemented");
+                    }
+
+                    grid.WorldPosition = gridDatum.Coordinates.Position;
+
+                    var modified = new List<(MapIndices position, Tile tile)>();
+                    foreach (var chunkData in gridDatum.ChunkData)
+                    {
+                        var chunk = grid.GetChunk(chunkData.Index);
+                        DebugTools.Assert(chunkData.TileData.Length == grid.ChunkSize * grid.ChunkSize);
+
+                        var counter = 0;
+                        for (ushort x = 0; x < grid.ChunkSize; x++)
+                        for (ushort y = 0; y < grid.ChunkSize; y++)
                         {
-                            chunk.SetTile(x, y, tile);
-                            modified.Add((new MapIndices(chunk.X * grid.ChunkSize + x, chunk.Y * grid.ChunkSize + y), tile));
+                            var tile = chunkData.TileData[counter++];
+                            if (chunk.GetTile(x, y).Tile != tile)
+                            {
+                                chunk.SetTile(x, y, tile);
+                                modified.Add((new MapIndices(chunk.X * grid.ChunkSize + x, chunk.Y * grid.ChunkSize + y), tile));
+                            }
                         }
+                    }
+
+                    if (modified.Count != 0)
+                    {
+                        GridChanged?.Invoke(this, new GridChangedEventArgs(grid, modified));
                     }
                 }
 
-                if (modified.Count != 0)
-                {
-                    GridChanged?.Invoke(this, new GridChangedEventArgs(grid, modified));
-                }
+                SuppressOnTileChanged = false;
             }
-
-            SuppressOnTileChanged = false;
         }
 
         public void ApplyGameStatePost(GameStateMapData data)
@@ -158,19 +178,25 @@ namespace Robust.Shared.Map
             if(data == null) // if there is no data, there is nothing to do!
                 return;
 
-            foreach (var grid in data.DeletedGrids)
+            if(data.DeletedGrids != null)
             {
-                if (_grids.ContainsKey(grid))
+                foreach (var grid in data.DeletedGrids)
                 {
-                    DeleteGrid(grid);
+                    if (_grids.ContainsKey(grid))
+                    {
+                        DeleteGrid(grid);
+                    }
                 }
             }
 
-            foreach (var map in data.DeletedMaps)
+            if(data.DeletedMaps != null)
             {
-                if (_maps.ContainsKey(map))
+                foreach (var map in data.DeletedMaps)
                 {
-                    DeleteMap(map);
+                    if (_maps.ContainsKey(map))
+                    {
+                        DeleteMap(map);
+                    }
                 }
             }
         }
